Rename files back to profile for compare changes more easily
